Allow additive operations.
Contacts / contact lists won't be deleted from Unifi Talk.

+++ talk/sync.py
@@ -7,7 +7,7 @@ logger = logging.getLogger(__name__)
 # CSV header for import to Unifi Talk.
 HEADER = 'first_name,last_name,company,job_title,email,mobile_number,home_number,work_number,fax_number,other_number'
 
-def sync_contacts(ctx, concatenate, grandstream, unifi_csv, unifi_talk):
+def sync_contacts(ctx, additive, concatenate, grandstream, unifi_csv, unifi_talk):
     contacts = GoogleContacts(ctx).parsed_contacts
     if len(contacts) == 0:
         logger.warning('no contacts to sync, no action taken')
@@ -19,7 +19,7 @@ def sync_contacts(ctx, concatenate, grandstream, unifi_csv, unifi_talk):
 
     if grandstream: write_grandstream_xml(ctx, concatenate, contacts)
     if unifi_csv: write_unifi_csv(ctx, concatenate, contacts)
-    if unifi_talk: sync_unifi_talk(ctx, concatenate, contacts)
+    if unifi_talk: sync_unifi_talk(ctx, additive, concatenate, contacts)
 
 
 def write_grandstream_xml(ctx, concatenate, contacts):
@@ -69,36 +69,45 @@ def write_unifi_csv_file(contacts, filename):
     logger.info(f'wrote {len(contacts)} contacts to {filename}')
 
 
-def sync_unifi_talk(ctx, concatenate, contacts):
+def sync_unifi_talk(ctx, additive, concatenate, contacts):
     api = TalkAPI(ctx)
 
     # Delete all contacts to ensure no incoming contacts have overlapping
     # phone numbers.
-    ids, response = api.delete_all_contacts()
-    if not response:
-        logger.error(f'failed to delete all contacts from Unifi Talk: {response}')
-        return
-    logger.info(f'deleted all {len(ids)} contacts from Unifi Talk') if ids is not None else logger.info('no contacts to delete from Unifi Talk')
+    if not additive:
+        ids, response = api.delete_all_contacts()
+        if not response:
+            logger.error(f'failed to delete all contacts from Unifi Talk: {response}')
+            return
+        if ids is not None:
+            logger.info(f'deleted all {len(ids)} contacts from Unifi Talk')
+        else:
+            logger.info('no contacts to delete from Unifi Talk')
     
     labels = ctx.obj['labels']
     concatenate_label = None
     if concatenate is not None:
         concatenate_label = [concatenate]
 
-    # We've deleted all contacts.  Let's also delete all contact lists that
-    # don't match the labels / concatenate label.  Don't touch the matching
-    # lists though - they may be associated with phones already (presumably
-    # by id) and recreating them could change their id and therefore associate
-    # the wrong contacts with a phone ... maybe.
-    deleted, failed = api.delete_all_contact_lists(concatenate_label or labels)
-    if len(failed) > 0:
-        logger.error(f'deleted {len(deleted)} but failed to delete {len(failed)} contact lists from Unifi Talk')
-        return
-    logger.info(f'deleted all {len(deleted)} contact lists from Unifi Talk') if len(deleted) > 0 else logger.info('no contact lists to delete from Unifi Talk')
+    if not additive:
+        # We've deleted all contacts.  Let's also delete all contact lists that
+        # don't match the labels / concatenate label.  Don't touch the matching
+        # lists though - they may be associated with phones already (presumably
+        # by id) and recreating them could change their id and therefore associate
+        # the wrong contacts with a phone ... maybe.
+        deleted, failed = api.delete_all_contact_lists(concatenate_label or labels)
+        if len(failed) > 0:
+            logger.error(f'deleted {len(deleted)} but failed to delete {len(failed)} unused contact lists from Unifi Talk')
+            return
+        if len(deleted) > 0:
+            logger.info(f'deleted {len(deleted)} unused contact lists from Unifi Talk')
+        else:
+            logger.info('no unused contact lists to delete from Unifi Talk')
 
     if not api.add_contact_lists_if_missing(concatenate_label or labels):
         logger.error(f'failed to create one or more contact lists {concatenate_label or labels}')
         return
+
     cl_map = api.get_contact_lists()
     if concatenate is not None:
         group_contacts = contacts.filter(labels).normalize(concatenate)
